class-4/pyproject.toml
[project]
name = "class-4"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "openai-agents>=0.2.1",
    "python-decouple>=3.8",
]
